chore: Move JSON backup files around
- The old `JSON.ts` file is moved to `JSON/v1.ts`, with a new `JSON/index.ts` file exporting the used hook.
- The `useExportBackup()` hook will use the export function from the latest implementation.
- We plan on having the `useImportBackup` hook figure out which version of the import function should be used.
